Setting up tests so I can test-drive a cleaner "SkyUnit2" interface
Many of the changes are coming so that we can write tests for SkyUnit *in SkyUnit*

One of the features that'll be added is a concept of multiple "instances" of the SkyUnit dataset. This is required to write SkyUnit tests in SkyUnit.

A side-effect should be better organized, easier to integrate with, and probably optimized code / interface.